move AbstractViewListener to sessions
