+++ plugin/core/windows.py
@@ -13,6 +13,7 @@ from .protocol import DiagnosticSeverity
 from .protocol import DocumentUri
 from .protocol import Error
 from .protocol import Location
+from .sessions import AbstractViewListener
 from .sessions import get_plugin
 from .sessions import Logger
 from .sessions import Manager
@@ -30,8 +31,6 @@ from .views import format_diagnostic_for_panel
 from .views import make_link
 from .workspace import ProjectFolders
 from .workspace import sorted_workspace_folders
-from abc import ABCMeta
-from abc import abstractmethod
 from collections import OrderedDict
 from collections import deque
 from subprocess import CalledProcessError
@@ -46,91 +45,6 @@ import urllib.parse
 
 
 _NO_DIAGNOSTICS_PLACEHOLDER = "  No diagnostics. Well done!"
-
-
-class AbstractViewListener(metaclass=ABCMeta):
-
-    TOTAL_ERRORS_AND_WARNINGS_STATUS_KEY = "lsp_total_errors_and_warnings"
-
-    view = None  # type: sublime.View
-
-    @abstractmethod
-    def session_async(self, capability_path: str, point: Optional[int] = None) -> Optional[Session]:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def sessions_async(self, capability_path: Optional[str] = None) -> Generator[Session, None, None]:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def session_views_async(self) -> Iterable[SessionViewProtocol]:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def on_session_initialized_async(self, session: Session) -> None:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def on_session_shutdown_async(self, session: Session) -> None:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def diagnostics_async(self) -> Iterable[Tuple[SessionBufferProtocol, Sequence[Tuple[Diagnostic, sublime.Region]]]]:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def diagnostics_intersecting_region_async(
-        self,
-        region: sublime.Region
-    ) -> Tuple[Sequence[Tuple[SessionBufferProtocol, Sequence[Diagnostic]]], sublime.Region]:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def diagnostics_touching_point_async(
-        self,
-        pt: int,
-        max_diagnostic_severity_level: int = DiagnosticSeverity.Hint
-    ) -> Tuple[Sequence[Tuple[SessionBufferProtocol, Sequence[Diagnostic]]], sublime.Region]:
-        raise NotImplementedError()
-
-    def diagnostics_intersecting_async(
-        self,
-        region_or_point: Union[sublime.Region, int]
-    ) -> Tuple[Sequence[Tuple[SessionBufferProtocol, Sequence[Diagnostic]]], sublime.Region]:
-        if isinstance(region_or_point, int):
-            return self.diagnostics_touching_point_async(region_or_point)
-        elif region_or_point.empty():
-            return self.diagnostics_touching_point_async(region_or_point.a)
-        else:
-            return self.diagnostics_intersecting_region_async(region_or_point)
-
-    @abstractmethod
-    def on_diagnostics_updated_async(self) -> None:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def on_code_lens_capability_registered_async(self) -> None:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def get_language_id(self) -> str:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def get_uri(self) -> str:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def do_signature_help_async(self, manual: bool) -> None:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def navigate_signature_help(self, forward: bool) -> None:
-        raise NotImplementedError()
-
-    @abstractmethod
-    def on_post_move_window_async(self) -> None:
-        raise NotImplementedError()
 
 
 def extract_message(params: Any) -> str:
